% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_indicators.R
\name{calculate_gross_output_identity}
\alias{calculate_gross_output_identity}
\title{Gross Output Identity Check (X = B * Y)}
\usage{
calculate_gross_output_identity(mrio_panel, year)
}
\arguments{
\item{mrio_panel}{MRIO data object.}

\item{year}{Year to analyze.}
}
\value{
Data.table comparing calculated vs observed Output.
}
\description{
Calculates Global Gross Output using the Global Leontief Inverse (B).
Checks the identity X = B * Y.
}
