% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_indicators.R
\name{calculate_total_exports_eq5}
\alias{calculate_total_exports_eq5}
\title{Calculate Total Bilateral Gross Exports}
\usage{
calculate_total_exports_eq5(mrio_panel, year)
}
\arguments{
\item{mrio_panel}{MRIO data object.}

\item{year}{Year to analyze.}
}
\value{
A \code{data.table} of total exports by sector.
}
\description{
Calculates total gross exports from Country S to the world based on the formula:
\eqn{EX_{sr} = A_{sr}X_r + Y_{sr}}.
}
