% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{get_sector_to_sector_matrix}
\alias{get_sector_to_sector_matrix}
\title{Extract Sector-to-Sector Trade Matrix (Z)}
\usage{
get_sector_to_sector_matrix(mrio_panel, year, exporter, importer)
}
\arguments{
\item{mrio_panel}{The MRIO data object.}

\item{year}{Integer. The year to analyze.}

\item{exporter}{Source country. Numeric Index (e.g., 8) OR Country Code (e.g., "PRC").}

\item{importer}{Destination country. Numeric Index (e.g., 43) OR Country Code (e.g., "USA").}
}
\value{
A sparse matrix (35 x 35) with row names formatted as \code{"Country_Sector"}.
Values represent flows in millions of USD (constant prices).
}
\description{
Extracts the specific N x N block of the intermediate transaction matrix ($Z$)
representing direct flows from the Exporter to the Importer.
}
