% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_indicators.R
\name{calculate_total_exports_eq5}
\alias{calculate_total_exports_eq5}
\title{Calculate Total Bilateral Exports (Equation 5)}
\usage{
calculate_total_exports_eq5(mrio_panel, year)
}
\arguments{
\item{mrio_panel}{MRIO data object.}

\item{year}{Year to analyze.}
}
\value{
Data.table of total exports by sector.
}
\description{
Calculates EX_sr = A_sr * X_r + Y_sr.
Returns the total exports of country s to the world (excluding domestic).
}
