% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition_bilateral.R
\name{compute_bilateral_wwz}
\alias{compute_bilateral_wwz}
\title{Detailed Bilateral WWZ Decomposition}
\usage{
compute_bilateral_wwz(mrio_panel, year, s_idx, r_idx)
}
\arguments{
\item{mrio_panel}{MRIO data object.}

\item{year}{Year to analyze.}

\item{s_idx}{Exporter index.}

\item{r_idx}{Importer index.}
}
\value{
Data.table of bilateral terms.
}
\description{
Performs bilateral trade decomposition (s -> r) using the WWZ framework.
}
