% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{get_sector_to_sector_matrix}
\alias{get_sector_to_sector_matrix}
\title{Extract Sector-to-Sector Trade Matrix}
\usage{
get_sector_to_sector_matrix(mrio_panel, year, exporter, importer)
}
\arguments{
\item{mrio_panel}{MRIO data object.}

\item{year}{Year to analyze.}

\item{exporter}{Exporter: Numeric Index (e.g. 8) OR Country Code (e.g. "PRC").}

\item{importer}{Importer: Numeric Index (e.g. 43) OR Country Code (e.g. "USA").}
}
\value{
35x35 Matrix.
}
\description{
Returns the N x N matrix of intermediate flows.
}
